#15 Update MultiFrameViewExtended's Method
So, this is very interesting. I have just added single line in the ApplyFormat method! And that is because, data coming from the ACCESS tables are not in required format, and due to fixed ACCESS data types, some field value came with extra blank spaces, so to remove those blank spaces, I have introduce TRIM function. That's All.