#15 Update PriceFormModel
Very interesting update as well. So, fields which are not-mandatory, are causing an issue because if field won't be change then validation won't be trigger to verify the changes and hence while clicking on the any CRUD buttons, it shows that, the field is empty! and for non-mandatory fields the check should be skipped but it wasn't so I has to explicitly set CheckField value for non-mandatory field to TRUE so, it won't affect during the final validation. Yes, But when user enter some values and then again make the field blank, at that time, individual check will handle this and will make CheckField Value to true! But it won't trigger if the user will not enter in this field. So this has to setup to handle validation issue.