#15 Update RepositoryExcel
Due to evolve of Application by integrating ACCESS repository, few methods in IServices has to be changed as well and IServices classes are being shared to all the repositories. So to keep every other repositories in align with IRespository methods, We have to update this class with few mandatory methods! And I have also updated Error handlings of methods.