#69 Update ExportFormModel Class
I have removed the lines of code, from the Export Report Procedure, because that step is exporting data without any proper format! Basically, when there is not criteria from the export form, we just simply export all the data but it also does not format columns where we want to make data in text format strictly so excel won't change it automatically to numbers.

e.g. Distribution channel does contain a string "01" but excel keep reformatting it to 1 which is invalid value