#77 Update modMain's Calling procedure
So, If error occurs during VIEW initialization then splash screen keep hanging on the screen and that is not correct behavior so, What happened was, when error occurs, code execution gets stopped and jump to the clean Exit label and there we DON'T have anything that can close or hide the splash screen. So I mentioned the Splash Screen Exit commands in the Clean Exit Too! So that will resolved the issue.