#90 Update AppModel Class
Ahh, what a finding. Basically, I was looking for the last row ID number and then I was adding 1 to it and creating New ID. But that approach goes wrong when PostgreSQL comes in the playground. Because when we update record in PostgreSQL table, it delete the target row first and then add updated row in the end of the table! So, that's why last row becomes false value! Because we could update any row. So then I have to introduce function that can return MAXIMUM number from the ID Column. That's what this is all about.