#87 Update modGlobal module
I searched about this bug, and found that, Application.Wait is only available with Excel Library and when we use Application Object without any context of library then it basically captures first Application object item from the list. So that might goes wrong when other office applications are open as well. So to avoid that mess, we are explicitly using method via library context of Excel.