#92 Update View
This is interesting. So, we are here hydrating the PriceFormModel property. But why? If we don't do this then it will be shown as REJECTED record only but we should show it as PENDING record because we want Approver to pick it and hence allowing approver to reevaluate the edited record.