#95  Update DataFormModel
Added Users Data Table for lookup. Unfortunately, we have to recreate table if we want to add any customize column like User_Name instead of User_ID

We unknowingly didn't used Init Method throughout for the Data Form Model, so I deleted it and also we didn't used Validator Services, so we removed that one too.