#95 Update Main View
We do not want users to edit such row which is empty or is heading.

We do not want any approver to delete user IDs 100 and 101 which is Kamal and Zoltan.

I changed the logic here because, earlier it would protecting only for the first two rows of the list box but if we edit any record then it shift in PostgreSQL hence that logic goes false. So to prevent that, I have finally explicitly mentioned the User ids those has to be prevent by any users to edit.