#110 Update AppModel Class
When we added new field as Email in the user manager, we actually missed to reflect the change in the password manager for a client. So to balance that I have added ActiveUserEmail Property in the AppModel Class. So When a client update his/her password, it doesn't produce any runtime errors. In Short one more future bug is resolved with this changes in this file.