#110 Update modGloval Module
So, with this issue, we should be able to delete user without setting up the password. So to manage that, I have explicitly flag password field as TRUE when operation is DELETE. because till now, we have only two options which is NEW and UPDATE but UPDATE also contain delete operation as well. So to differentiate the operation, we have to add new enum in the FormOperation ENUMS defination.